// src/api/generated/model/competitionRewardRuleDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */
import type { CompetitionRewardRuleDtoType } from './competitionRewardRuleDtoType';
import type { CompetitionRulePositionDto } from './competitionRulePositionDto';

export interface CompetitionRewardRuleDto {
  /** Unique identifier */
  id: string;
  /** Timestamp of when the record was created */
  createdAt: string;
  /** Timestamp of the last update of the record */
  updatedAt: string;
  /** Тип распределения награды */
  type: CompetitionRewardRuleDtoType;
  /**
   * Начало диапазона мест
   * @nullable
   */
  rankFrom: number | null;
  /**
   * Конец диапазона мест
   * @nullable
   */
  rankTo: number | null;
  /**
   * Минимальный порог баллов
   * @nullable
   */
  minPoints: number | null;
  /**
   * Спринт; null для правила события
   * @nullable
   */
  sprintId: string | null;
  /**
   * Событие; null для правила спринта
   * @nullable
   */
  eventId: string | null;
  rewards: CompetitionRulePositionDto[];
}